Extract chart in new component
