+++ packages/client/src/modules/charts/index.ts
@@ -1,0 +1,3 @@
+import { StackedEnergyBars } from "./StackedEnergyBars";
+
+export { StackedEnergyBars };
